requirements.txt
Django==1.8.4
django-braces==1.8.1
django-cors-headers==1.1.0
django-oauth-toolkit==0.9.0
djangorestframework==3.0.4
oauthlib==1.0.1
Pillow==2.9.0
six==1.9.0
